refactor(gamemodes): make gamemode module contain .cpp files
before it was headers only
