+++ inexor/client/CMakeLists.txt
@@ -72,3 +72,4 @@ require_util(${CLIENT_BINARY})
 require_ui(${CLIENT_BINARY})
 require_texture(${CLIENT_BINARY})
 require_filesystem(${CLIENT_BINARY})
+require_gamemode(${CLIENT_BINARY})
