+++ inexor/server/CMakeLists.txt
@@ -27,3 +27,4 @@ require_network(${SERVER_BINARY} "SERVER STANDALONE SERVMODE")
 require_util(${SERVER_BINARY})
 require_crashreporter(${SERVER_BINARY})
 require_filesystem(${SERVER_BINARY})
+require_gamemode(${SERVER_BINARY})
